requirements.txt
aiohttp==3.9.5
pydantic==2.7.1
python-dotenv==1.0.1